reorganize editable text implementation for maintainability
