+++ crates/gpui_elements/src/editable_text.rs
@@ -2,6 +2,7 @@ pub mod actions;
 mod caret;
 mod element;
 mod history;
+mod layout;
 mod state;
 mod storage;
 
@@ -10,7 +11,6 @@ pub use state::*;
 pub use storage::*;
 
 /* TODO list
-- remove gpuikit based input
 - text sanitation
 - add page up/down actions to nav by an entire page or expand selection by an entire page
 - test IME (char palette only available on macos)
